Point addition complete

+++ eddsa/src/main/java/io/moatwel/crypto/eddsa/Curve.java
@@ -30,5 +30,5 @@ public interface Curve {
 
     Coordinate getD();
 
-    int getA();
+    BigInteger getA();
 }
